+++ examples/example-combine-pathf-metadata/build.rs
@@ -1,0 +1,3 @@
+fn main() {
+    mingling::build::analyze_and_build_type_mapping().unwrap();
+}