+++ mingling_core/src/asset.rs
@@ -5,6 +5,7 @@ pub(crate) mod enum_tag;
 pub(crate) mod global_resource;
 pub(crate) mod help;
 pub(crate) mod lazy_resource;
+pub(crate) mod metadata;
 pub(crate) mod node;
 pub(crate) mod renderer;
 pub(crate) mod routable;