+++ mingling_core/src/lib.rs
@@ -70,6 +70,7 @@ pub use crate::asset::enum_tag::*;
 pub use crate::asset::global_resource::*;
 pub use crate::asset::help::*;
 pub use crate::asset::lazy_resource::*;
+pub use crate::asset::metadata::*;
 pub use crate::asset::node::*;
 pub use crate::asset::renderer::*;
 pub use crate::asset::routable::*;