+++ mingling_macros/src/attr.rs
@@ -6,6 +6,7 @@ pub(crate) mod completion;
 #[cfg(feature = "clap")]
 pub(crate) mod dispatcher_clap;
 pub(crate) mod help;
+pub(crate) mod metadata;
 pub(crate) mod mlint;
 #[cfg(feature = "extras")]
 pub(crate) mod program_setup;