+++ mingling_pathf/src/pattern_analyzer.rs
@@ -36,6 +36,7 @@ pub fn init_with_config(config: PathfinderConfig) -> PatternAnalyzer {
     analyzer.add_pattern(CommandPattern);
     analyzer.add_pattern(RendererPattern);
     analyzer.add_pattern(HelpPattern);
+    analyzer.add_pattern(MetadataPattern);
     analyzer.add_pattern(CompletionPattern);
     analyzer.add_pattern(DispatcherPattern::new(config.use_dispatch_tree));
     analyzer.add_pattern(DispatcherClapPattern::new(config.use_dispatch_tree));