+++ mingling_pathf/src/patterns.rs
@@ -9,6 +9,7 @@ pub use dispatcher_clap::*;
 pub use group::*;
 pub use grouped_derive::*;
 pub use help::*;
+pub use metadata::*;
 pub use pack::*;
 pub use renderer::*;
 
@@ -21,5 +22,6 @@ mod dispatcher_clap;
 mod group;
 mod grouped_derive;
 mod help;
+mod metadata;
 mod pack;
 mod renderer;